Ajuste teste unitário bebidaService

+++ src/main/java/br/com/portifolio/ProjetoBebidas/model/domain/Bebida.java
@@ -1,26 +1,24 @@
 package br.com.portifolio.ProjetoBebidas.model.domain;
 
-import br.com.portifolio.ProjetoBebidas.model.Enum.TipoBebidaEnum;
-
 public class Bebida {
 
-    private Integer id;
+    private Long id;
     private String nome;
     private TipoBebida tipoBebida;
 
     public Bebida(){}
-    public Bebida(Integer id, String nome, TipoBebida tipoBebida) {
+    public Bebida(Long id, String nome, TipoBebida tipoBebida) {
         this.id = id;
         this.nome = nome;
         this.tipoBebida = tipoBebida;
         //this.bebidaSessoes = bebidaSessoes;
     }
     // Getters and Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
